Renamed StateChannel to ContextChannel and "anchor" to "state"
The new interface is more intuitive, for example:

class ChatRoomChannel(ContextChannel):
    class Meta:
        state = chat.serializers.ChatRoomNestedSerializer

Now we can conceptually support actions, as the channel is more than
just "State". The actions PR has been open for months waiting for this
resolution.

The "anchor" concept is still usually internally, it was just renamed
on the framework API.
