+++ rxdjango-react/src/ContextChannel.interfaces.ts
